% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apm-package.R
\docType{package}
\name{apm-package}
\alias{apm}
\alias{apm-package}
\title{apm: Averaged Prediction Models}
\description{
In panel data settings, specifies set of candidate models, fits them to data from pre-treatment validation periods, and selects model as average over candidate models, weighting each by posterior probability of being most robust given its differential average prediction errors in pre-treatment validation periods. Subsequent estimation and inference of causal effect's bounds accounts for both model and sampling uncertainty, and calculates the robustness changepoint value at which bounds go from excluding to including 0. Package also includes a range of diagnostic plots, such as those illustrating models' differential average prediction errors and the posterior distribution of which model is most robust.
}
\author{
\strong{Maintainer}: Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})

Authors:
\itemize{
  \item Thomas Leavitt \email{thomas.leavitt@baruch.cuny.edu} (\href{https://orcid.org/0000-0002-3668-6409}{ORCID})
  \item Laura Hatfield \email{hatfield@hcp.med.harvard.edu} (\href{https://orcid.org/0000-0003-0366-3929}{ORCID})
}

}
\keyword{internal}
