% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apm_error_handler}
\alias{apm_error_handler}
\title{Error handler for tryCatch calls.}
\usage{
apm_error_handler(e)
}
\arguments{
\item{e}{Error}
}
\value{
side effects
}
\description{
Error handler for tryCatch calls.
}
\note{
Needs global `APM_MODULE_NAME`.
}
\author{
Wayne Folta
}

