% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{close_db}
\alias{close_db}
\title{Close the DB}
\usage{
close_db(db_conn)
}
\arguments{
\item{db_conn}{the \code{db_conn} value returned from \code{initialize_db}}
}
\description{
\code{close_db} closes the connection to the NRDB.
}
\author{
Wayne Folta
}

