% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convTF012}
\alias{convTF012}
\title{Convert fields}
\usage{
convTF012(varTF012, input_type, output_type)
}
\arguments{
\item{varTF012}{The input}

\item{input_type}{Input type (ignored for now)}

\item{output_type}{Output type ("numeric", "factor1", "boolean")}
}
\value{
converted thingie
}
\description{
\code{convTF012} converts several varieties of input flag fields
to several kinds of output flag fields.
}
\details{
Inputs can be:
boolean (TRUE, FALSE, NA), boolean strings ("t", "T", "f", "F", NA),
boolean numbers (0, 1, 2, where 0 is FALSE, 1 is TRUE, and 2 is NA).

Outputs can be boolean (TRUE, FALSE, NA), numeric (0, 0.5, 1, where
0 if FALSE, 1 is TRUE and 0.5 is NA), or factor ("T", "F", "NA").

For example the SOM prefers numeric outputs, while the SVM prefers
factor. (Factor output also does not have NA's, which can cause
rows to be omitted in many models. What would be an NA is a factor
level "NA".)

Currently there is an input_type and an output_type, but input_type
is ignored since the currently accepted types work without special
case handling. Of course output_type is critical, since each type
has special case handling.
}
\author{
Wayne Folta
}

