% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_rows}
\alias{count_rows}
\title{Count rows in any table}
\usage{
count_rows(conn, tbl)
}
\arguments{
\item{conn}{Database connection handle.}

\item{tbl}{Database table name.}
}
\value{
An integer number of rows in the table.
}
\description{
Count rows in any table
}
\author{
Tom Shafer
}

