% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_begin}
\alias{db_begin}
\title{Set up Score Writing Transaction}
\usage{
db_begin(conn)
}
\arguments{
\item{conn}{the \code{conn} part of the \code{db_conn} value returned from \code{initialize_db}.}
}
\description{
\code{db_begin} starts a score transaction that can be committed or rolled back.
}
\author{
Wayne Folta
}

