% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_commit}
\alias{db_commit}
\title{Commit of Delete and Write Scores}
\usage{
db_commit(conn)
}
\arguments{
\item{conn}{the \code{conn} part of the \code{db_conn} value returned from \code{initialize_db}.}
}
\description{
\code{db_commit} commits score writing.
}
\author{
Wayne Folta
}

