% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_rollback}
\alias{db_rollback}
\title{Rollback of Delete and Write Scores}
\usage{
db_rollback(conn)
}
\arguments{
\item{conn}{the \code{conn} part of the \code{db_conn} value returned from \code{initialize_db}.}
}
\description{
\code{db_rollback} rolls back the score-writing transaction. Used in case of error.
}
\author{
Wayne Folta
}

