% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_scores}
\alias{delete_scores}
\title{Delete (Old) Scores}
\usage{
delete_scores(conn, score_cd, cycle_date, scoreTableName)
}
\arguments{
\item{conn}{the \code{conn} part of the \code{db_conn} value returned from \code{initialize_db}.}

\item{score_cd}{the the character string code of this score type.}

\item{cycle_date}{the cycle date (numeric YYYYQ) of the current cycle.}

\item{scoreTableName}{fully-qualified table from which to delete scores. Could be redirected}
}
\description{
\code{delete_scores} deletes old scores from this Cycle Date.
}
\details{
This code was modified so it's not an error to delete scores from a
non-existent score table. However, \code{write_scores} doesn't create
a table, so this turns out to be less than useful sometimes.
}
\author{
Wayne Folta
}

