% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance_score}
\alias{distance_score}
\title{Distance Risk Module}
\usage{
distance_score(cycle_date = NULL, scoreTableName = "nrd.enty_score",
  inputDF = NULL)
}
\arguments{
\item{cycle_date}{The cycle date being scored}

\item{db_schema}{Where to write scores to. Default is nrd.}
}
\value{
Returns 0 on success, >0 on failure, look at the \code{return}
statements to identify where the failure occurred.
}
\description{
This module is intended to replace the Location Risk Module used in Kansas
and Idaho. New York did not find the Location Risk Module helpful because
many of their claimants file legally from nearby states as well as Canada, 
while some places within the state may be high risk if it is far from their
home address.
}
\details{
The location of the home address is pulled from the zip code given in 
\code{nrd.addr}, which is then matched with a database containing zip codes
and corresponding latitude and longitude coordinates. The home address
location is then compared to the lat/long coordinates for the current 
certification's IP address to get the distance in miles. The zip code
database containing lat/lon coordinates is from federalgovernmentzipcodes.us
(version updated 1/22/2012). The current certification's IP address is
determined from \code{nrd.aggr_claimt_sessn}.

The base unit is a single certificaiton (one week) for a claimant, but some
cumulative information is collected for the entire claim. The final score
reported uses the average between two numbers:
  1. Distance from home as described above
  2. Average distance between all certifications in the claim
and is scaled from 0 to 100.
}
\author{
Daniel Brannock
}

