% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eepd-package.R
\docType{package}
\name{eepd-package}
\alias{eepd}
\alias{eepd-package}
\title{eepd: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})

Authors:
\itemize{
  \item Thomas Leavitt \email{thomas.leavitt@baruch.cuny.edu} (\href{https://orcid.org/0000-0002-3668-6409}{ORCID})
  \item Laura Hatfield \email{hatfield@hcp.med.harvard.edu} (\href{https://orcid.org/0000-0003-0366-3929}{ORCID})
}

}
\keyword{internal}
