% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excl_tables}
\alias{excl_tables}
\title{Form an SQL string excluding tables of claimants from a query}
\usage{
excl_tables(the_list = NULL)
}
\arguments{
\item{the_list}{List of tables to exclude.}
}
\value{
SQL string of exclusion clauses.
}
\description{
This returns something like:
   and claimt_id not in (select distinct claimt_id from A)
   and claimt_id not in (select distinct claimt_id from B)
   ...
}
\author{
Tom Shafer
}

