% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_param_from_nrdb}
\alias{fetch_param_from_nrdb}
\title{Fetch a model parameter from the NRDB}
\usage{
fetch_param_from_nrdb(con, param)
}
\arguments{
\item{con}{Database connection.}

\item{param}{Parameter name in \code{ref.score}.}
}
\value{
Parameter or \code{NULL}.
}
\description{
Fetch a model parameter from the NRDB
}

