% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_git_version}
\alias{find_git_version}
\title{Return a string with the current git status.}
\usage{
find_git_version()
}
\value{
String, e.g., "1.01-90-g8c319fe".
}
\description{
Return a string with the current git status.
}
\author{
Tom Shafer
}

