% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_custIdList}
\alias{get_custIdList}
\title{Input the Customer ID List}
\usage{
get_custIdList(conn, train.start)
}
\arguments{
\item{conn}{Connection returned from \code{initialize_db}.}

\item{train.start}{the earliest year/quarter (numeric YYYYQ) to be read in.}
}
\value{
converted thingie
}
\description{
\code{get_custIdList} reads in the customer ID list from \code{nrd.aggr_claimt_benf}.
}
\author{
Wayne Folta
}

