% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_emp_edgelist}
\alias{get_emp_edgelist}
\title{Read in employer data to create an edge list}
\usage{
get_emp_edgelist(db_conn, sql_quarter)
}
\arguments{
\item{db_conn}{Database connection}

\item{sql_quarter}{Quarter to anchor on for the edgelist}
}
\value{
side effects
}
\description{
Read in employer data to create an edge list
}
\author{
Micah Darrell
}

