% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_version}
\alias{get_model_version}
\title{Return the model version stored in the APM package}
\usage{
get_model_version()
}
\value{
Character vector
}
\description{
Return the model version stored in the APM package
}
\author{
Tom Shafer
}

