% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pii_edgelist}
\alias{get_pii_edgelist}
\title{Read in claimant data to create an edge list}
\usage{
get_pii_edgelist(db_conn)
}
\value{
side effects
}
\description{
Read in claimant data to create an edge list
}
\author{
Micah Darrell
}

