% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_model_vars}
\alias{impute_model_vars}
\title{Impute numeric variables for statistical models}
\usage{
impute_model_vars(df, impute.list)
}
\arguments{
\item{df}{Data frame (or data table, we convert here)}

\item{impute.list}{List(col.name=impute.value) of columns and their
imputation values. If impute.value=NA, this is treated as if the column
were not specified at all.}
}
\value{
Data frame in which requested NAs are replaced with imputed values.
}
\description{
For each column in the data frame: if it should be imputed (if it appears in
impute.list and is not NA there), replace all NA values with the specified
value from impute.list. Otherwise, just return the original column.
}
\author{
Tom Shafer
}

