% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_values_dt}
\alias{impute_values_dt}
\title{Impute values by reference within a data table}
\usage{
impute_values_dt(dt, impute.list)
}
\arguments{
\item{dt}{A data table.}

\item{impute.list}{A list for which keys are column names in the data table
and values are their imputed values.}
}
\value{
Returns an object with "error" in \code{class(obj)} if
        \code{tryCatch} detects a problem, otherwise a list of messages.
}
\description{
For each key in the list, use data table for imputation. This works by
selecting the correct row (\code{dt[is.na(dt[[key]]), ...])} and then
imputing via data.table's \code{ov[, c(<chr>) := <...>]} syntax.
}
\author{
Tom Shafer
}

