% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_apm}
\alias{initialize_apm}
\title{Initialize the APM utilities}
\usage{
initialize_apm(module_name)
}
\arguments{
\item{module_name}{the module name to be set (globally) for error and logging}
}
\value{
converted thingie
}
\description{
\code{initialize_apm} initializes the utilities, including checking for the
existence of \code{APM_} variables that indicate we're called from Kettle, and
setting the \code{APM_MODULE_NAME} that is used for errors and logging.
}
\note{
If the global variable \code{APM_DB_NAME} exists, we've been called from
Kettle, which sets it (and other \code{APM_} variables). If it does not exist,
get the Windows System Environment variable \code{APM_WORKING} and use that to
find \code{nrdbPassword.R} and source it which will set the variables.
}
\author{
Wayne Folta
}

