% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_db}
\alias{initialize_db}
\title{Initialize the DB}
\usage{
initialize_db()
}
\value{
conn the \code{dbConnect} returned value

DRV the \code{JDBC} returned value
}
\description{
\code{initialize_db} initializes the connection to the NRDB.
}
\author{
Wayne Folta
}

