% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_scoring.R
\name{ip_location_score}
\alias{ip_location_score}
\title{IP Location Score}
\usage{
ip_location_score(cycle_date = NULL, db_schema = "nrd",
  drop_tmp_tables = TRUE, exclude_cellular = FALSE)
}
\arguments{
\item{cycle_date}{The APM cycle date.}

\item{db_schema}{The database schema to write scores and related tables to.
Default: "nrd".}

\item{drop_tmp_tables}{Remove created TMP tables.}

\item{exclude_cellular}{Whether or not the module should exclude
records identified as originating from cellular IPs.}
}
\value{
An integer status code, like other models.
}
\description{
This module reads in the ip base table, looks at the ip location flags and
assigns a location score for each claim using the following methodology:
}
\details{
- If someone filed a claim from a different state/province than was listed
  in their file and they're outside US OR they used an anonymous proxy (at.
  least once in the past "x" days): 100
- If someone filed a claim from a different state/province than was listed
  in their file and they're inside the US but outside the the state and
  neighboring states (at least once in the past "x" days): 50
- If someone filed a claim from a different state/province than was listed
  in their file and they're inside the US AND in the local area (all the time
  in the past "x" days): 25
- If no state of record is associated with claimant: 10
- If they've always filed from the same state as was listed in their file
  (in the past "x" days): 0 (no scores of 0 written to final score table)

Assumption(s):
- Those serving in the armed forces abroad (state code = 'AE' or 'AP') are
  excluded

The following parameters are pulled from a parameter table to run this model:
- ${APM_CYCL_DT}: Kettle date insert (also set via function arg)
}
\examples{
\dontrun{
rc <- ip_location_score()
}

}

