% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{knn.storage}
\alias{knn.storage}
\title{K nearest neighbors model object}
\format{See training/scoring functions.}
\usage{
knn.storage
}
\description{
\strong{This model object should never be put into version control.}
}
\keyword{datasets}

