% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_sample_table_with_n_claimts}
\alias{make_sample_table_with_n_claimts}
\title{Draw all records from nrd.aggr_claimt_benf for a random sample of claimants}
\usage{
make_sample_table_with_n_claimts(conn, start_qtr, end_qtr, n_claimts, tbl_name,
  exclude_claimts = NULL, target_cols = c("ovpaymt_amt"))
}
\arguments{
\item{conn}{Database connection handle.}

\item{start_qtr}{Start quarter number.}

\item{end_qtr}{End quarter number (inclusive).}

\item{n_claimts}{Number of claimants to sample.}

\item{tbl_name}{Name (with schema) of the table to create.}

\item{exclude_claimts}{Character vector of tables whose claimants should not
be included.}

\item{target_cols}{Character vector of columns to join from the
overpayment table.}
}
\value{
Integer with the number of rows in the specified table.
}
\description{
Sample nrd.aggr_claimt_benf, possibly excluding claimants from a list of
tables, and left join with the benefits table to form an ABT.
}
\author{
Tom Shafer
}

