% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_cap}
\alias{max_cap}
\title{Transformation function to control max value for a variable}
\usage{
max_cap(var, cap)
}
\arguments{
\item{var}{Input to be capped}

\item{cap}{Value to be capped at}
}
\value{
The capped variable
}
\description{
Caps variables at a given maximum. Necessary because some payments are
duplicated without a good way to filter out the bad ones. But no person was
legitimately paid more than the maximum payment amount.
}
\author{
Daniel Brannock
}

