% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_log}
\alias{min_log}
\title{Transformation function to take a log}
\usage{
min_log(var, minval)
}
\arguments{
\item{var}{Input to be logged}

\item{minval}{Minimum allowed value}
}
\value{
The capped variable
}
\description{
This function does two things.
  1. Ensures that the minimum value for a variable is not above a predefined
     minimum. This is critical for model robustness, since new unexpected
     negative values would otherwise break the log function (we've seen
     these in NY).
  2. Brings all values above zero to allow them to be logged.
}
\author{
Daniel Brannock
}

