% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{model.version}
\alias{model.version}
\title{Model version character vector}
\format{A character vector naming the particular model object.}
\usage{
model.version
}
\description{
Model version character vector
}
\keyword{datasets}

