% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_check_and_omit}
\alias{na_check_and_omit}
\title{Report and scrub rows with NA values}
\usage{
na_check_and_omit(ov)
}
\arguments{
\item{ov}{Benefits + overpayments (usually a data table or data frame)}
}
\value{
Object stripped of rows containing NA values
}
\description{
Report and scrub rows with NA values
}
\author{
Wayne Folta, Stuart Price; refactored by Tom Shafer
}

