% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_scoring.R
\name{network_score}
\alias{network_score}
\title{PII and Employment Network Scoring}
\usage{
network_score(cycle_date = NULL, prevDays = 90,
  scoreTableName = "nrd.enty_score", emplrQuarters = 2, piiYears = 5,
  ipWeeks = 13)
}
\arguments{
\item{cycle_date}{Weekend date to calculate scores up to}

\item{prevDays}{Days prior to the cycle_date to include for scoring}

\item{scoreTableName}{Postgres table in which to put scores; set NULL to return a data frame}

\item{emplrQuarters}{Number of quarters to seek back for employment history}

\item{piiYears}{Number of years to seek back for PII connections}

\item{ipWeeks}{Number of weeks to seek back for IP connections}
}
\value{
Integer code to Kettle or data frame if `scoreTableName = NULL`
}
\description{
This module connects claimants to others identified as having commited fraud
via three factors:
}
\details{
1. Employment history
2. PII
3. IP address sharing

Because of time and testing constraints (cannot easily test models that use
aggregated data, as the data is not available retroactively), all facets
of the model are reimplemented in R.

Based on initial testing with 2015 NY data (see Rmd file), the model provides
lift by summing up claimants' ranks in the subscores (i.e., average rank). A
more probabilistic model is possible (like the 1st version), but needs to
be able to combine subscores in a logical way.
}
\author{
Micah Darrell

3-3-2016 Kenny Darrell performace tweaks

4-12-2016 Wayne Folta updates to latest APM

11-2-2016 Tom Shafer rewrites for better performance with a large network
}

