% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_scoring.R
\name{new_hire_score}
\alias{new_hire_score}
\title{New Hire Scoring}
\usage{
new_hire_score(cycle_date = NULL, db_schema = "nrd")
}
\arguments{
\item{cycle_date}{The APM cycle date.}

\item{db_schema}{The database schema to write scores and related tables to.
Default: "nrd".}
}
\value{
An integer status code, like other models.
}
\description{
This function checks for un/under reported income from claimants. It reads
in the new hire table (NH_EMPLOYEE) and checks it against the list of
claimants who received benefits in the @time_window following their start
date. Not all claims will be fraudulent, since underemployment is still an
issue.
}
\details{
A claimant's risk will decrease with respect to their reported income
relative to their total benefits. Since the amount a claimant can earn
before losing benefits changes state to state, a variable @wage_threshold
is included.

Variables/Parameters:
- starts_after:        Gives lower bound to search window for new hire list
- time_window:         The length of time from start date of the new hire to
                       check for benefits being awarded.
- wage_threshold_low:  The fraction of WBA a claimant can earn before
                       benefits are reduced.
- wage_threshold_high: The fraction of WBA a claimant can earn before not
                       receiving benefits.

\code{act_wage_threshold_high = total_wba * wage_threshold_high}
\code{act_wage_threshold_low  = total_wba * wage_threshold_low}

Score:
\code{
if(claimed_earnings > act_wage_threshold_low):
	 100 - 100*(claimed_earnings/act_wage_threshold_high)
else:
	 100
}

The following parameters are pulled from a parameter table to run this model:
- ${APM_CYCL_DT}: Kettle date insert (or passed as function arg)
}
\examples{
rc <- new_hire_score()

}

