% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ny_extensions.R
\name{ny_qwr_score}
\alias{ny_qwr_score}
\title{NYSDOL QWR Cross-check}
\usage{
ny_qwr_score(cycle_date = NULL, prev_days = 90,
  score_table_name = "nrd.enty_score")
}
\arguments{
\item{cycle_date}{APM cycle date as a character vector.}

\item{prev_days}{Claimants who hav certified within `prev_days` days will
be assigned a score.}

\item{score_table_name}{NRDB score table name or \code{NULL}.
If \code{NULL}, return a data frame instead of writing to the NRDB.}
}
\value{
An integer return code (or a data frame containing the scores if
  \code{score_table_name} is \code{NULL}).
}
\description{
Implement NYSDOL's version of the QWR cross-check based on a
\code{STORED_VALUE} that estimates a claimant's wages in a quarter.
}
\examples{
\dontrun{
library(APM)
qwr <- ny_qwr_score()
}
}

