% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ny_extensions.R
\name{ny_statistical_model}
\alias{ny_statistical_model}
\title{NYSDOL Statistical Model Wrapper}
\usage{
ny_statistical_model(model, effect = 0.75, cycle_date = NULL,
  prevDays = 90, inputDF = NULL, scoreTableName = "nrd.enty_score",
  k = NULL)
}
\arguments{
\item{model}{DAPM module code as a character vector (`"KNN_KNN"`, `"RF_RF"`,
or `"SVM_SVM"`)}

\item{effect}{Scale the usual risk score by `(1 - effect * nwnr_prob)`}

\item{cycle_date}{APM cycle date as a character vector}

\item{prevDays}{Claimants who have certified within `prevDays` days will
be assigned a score}

\item{inputDF}{A data frame of predictors, if a pull from the NRDB is not
desired}

\item{scoreTableName}{NRDB score table name or `NULL`
If `NULL`, return a data frame instead of writing to the NRDB.}

\item{k}{Passed to \link{k_nearest_neighbors_score} if `model` = "KNN_KNN"}
}
\value{
An integer return code (or a data frame containing the scores if
  `scoreTableName` is `NULL`).
}
\description{
A wrapper function that applies a by-hand Bayesian Rule List model
to identify likely non-willful non-recoverable (NWNR) overpayments. This
corrects an issue with the NRDB's targets whereby the modules pinpoint
overpayments that NY cannot recover.

The scores produced by the statistical models are downweighted proportionally
to the BRL model's posterior probability of being NWNR.

This function depends on the "Suggested" package `sbrl` and an internal
not-documented data set `brl.storage`.
}

