% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{output_message}
\alias{output_message}
\title{Output (log) messages}
\usage{
output_message(mess, doMemory = FALSE, clear = FALSE)
}
\arguments{
\item{mess}{string message to output}

\item{doMemory}{add the amount of RAM available to R}

\item{clear}{clear the screen (output ^L) before message}
}
\description{
\code{output_message} outputs a (log) message to stdout, optionally including
the amount of allocated RAM and optionally clearing the screen.
}
\author{
Wayne Folta
}

