% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_scoring.R
\name{qwr_score}
\alias{qwr_score}
\title{QWR2 Score}
\usage{
qwr_score(cycle_date = NULL, db_schema = "nrd")
}
\arguments{
\item{cycle_date}{The APM cycle date.}

\item{db_schema}{The database schema to write scores and related tables to.
Default: "nrd".}
}
\value{
Integer status code, like other models.
}
\description{
Create the QWR+ score.
04 Aug 2015
}
\details{
This metric is aimed at folks who continue to collect benefits once they
start a new job, or otherwise are double-dipping.

It gives no score for a quarter in which a person is not found in the QWR reports.

The f1 and f2 components are looking for a large gap between (employer-reported) wages
and (claimant-reported) earnings: both in dollar terms and also as a percentage.

The f3 and f4 components are looking for claimants who claimed for a large percentage
of the quarter and of their possible benefits. And the f5 component is a bonus for
folks reporting a high percentage of benefit-awarding weeks with zero earnings.

USES:     benefits, quarterly_wage

CREATES:  qwr_score
}
\examples{
\dontrun{
rc <- qwr_score()
}

}
\author{
Wayne Folta
}

