% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{random_forest_train}
\alias{random_forest_train}
\title{Train a DAPM Random Forest Model}
\usage{
random_forest_train(df, transform.funcs = list(), impute.vals = list(),
  params = list(num_trees = 500))
}
\arguments{
\item{df}{A not-transformed data frame of predictors *plus* an additional
boolean column named \code{target}.}

\item{transform.funcs}{A list of APM-available functions that will be applied
to the input data both for training and for scoring. No transformation
will be applied to columns that do not have corresponding keys
in this list.}

\item{impute.vals}{A list of values that will be imputed to missing
observations in both the training and scoring data. No imputation will
be performed for columns that do not have corresponding keys.}

\item{params}{A list of parameters available for use by the model. These will
persist with the model object.}
}
\value{
Returns 0 on success and an integer > 0 if an error occured. These
        integer codes should be unique.
}
\description{
This is a simpler, more opinionated training function than has been included
in DAPM previously, but this routine will extend more easily to multiple
states with different data.
}
\details{
Assumptions:

1. The input data \code{df} is already balanced with respect to the target.
2. The input data \code{df} contains *only* the (a) target and (b) features
   used to predict the target. These columns will be the only ones used in
   the final model.
}
\author{
Tom Shafer
}

