% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_overpayments}
\alias{read_overpayments}
\title{Read Overpayments}
\usage{
read_overpayments(conn)
}
\arguments{
\item{conn}{the \code{conn} part of the \code{db_conn} value returned from \code{initialize_db}.}
}
\description{
\code{read_overpayments} reads in the entire overpayments table (\code{nrd.aggr_claimt_ovpaymt}).
}
\author{
Wayne Folta
}

