% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{rf.storage}
\alias{rf.storage}
\title{Random forest model object}
\format{See training/scoring functions.}
\usage{
rf.storage
}
\description{
\strong{This model object should never be put into version control.}
}
\keyword{datasets}

