% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_param}
\alias{set_param}
\title{Assign a parameter or optionally fill from NRDB}
\usage{
set_param(param, con, input = NULL)
}
\arguments{
\item{param}{Parameter name in the NRDB.}

\item{con}{NRDB connection object.}

\item{input}{User-assigned input. Non-\code{NULL} inputs short circuit
NRDB assignment.}
}
\value{
A parameter.
}
\description{
Assign a parameter or optionally fill from NRDB
}

