% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_param_lgl}
\alias{set_param_lgl}
\title{Assign a parameter or optionally fill from NRDB}
\usage{
set_param_lgl(...)
}
\arguments{
\item{...}{Arguments to pass to \code{set_param}.}
}
\value{
A boolean.
}
\description{
Assign a parameter or optionally fill from NRDB
}
\examples{
\dontrun{
j <- set_param_lgl(param = "EXCLUDE_CELLULAR", con = db_conn$conn)
}
}

