% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_param_num}
\alias{set_param_num}
\title{Assign a parameter or optionally fill from NRDB}
\usage{
set_param_num(...)
}
\arguments{
\item{...}{Arguments to pass to \code{set_param}.}
}
\value{
A numeric.
}
\description{
Assign a parameter or optionally fill from NRDB
}
\examples{
\dontrun{
i <- set_param_num(param = "K", con = db_conn$conn, input = k)
}
}

