% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_scoring.R
\name{sharing_ip_score}
\alias{sharing_ip_score}
\title{Sharing IP Scoring}
\usage{
sharing_ip_score(cycle_date = NULL, db_schema = "nrd",
  drop_tmp_tables = TRUE, exclude_cellular = FALSE)
}
\arguments{
\item{cycle_date}{The APM cycle date.}

\item{db_schema}{The database schema to write scores and related tables
to. Default: "nrd".}

\item{drop_tmp_tables}{Should the routine remove temporary tables created
along the way to risk scores? Default: TRUE.}

\item{exclude_cellular}{Whether or not the module should exclude
records identified as originating from cellular IPs.}
}
\value{
Integer status code, like other models.
}
\description{
This module reads in the IP base table, determines which users have have
frequently shared IP addresses over a certain time period and inserts risk
scores into the entity score table indicating the severity of the sharing
behavior for the most recent bus_perd_end_dt.  The (quantile) scores are
calculated using ip sharing behavior from the entire date range, but only
the last bus_perd_end_dt is kept.
}
\details{
The following variables/parameters are set at the beginning of the script:
- IP_MIN_USERS: The minimum # of users sharing an IP address within the same
                week required for those to be eligible for a risk score.
- IP_DECAY_FACTOR: The factor to be used for a time decay of scores (in
                   e^(-kt), k = decay factor)

NOTE: This module also uses IP location to complement its score. If a user's
ip is shared and it's coming from outside the states listed in the
claimant's file, it's more risky.
}
\examples{
\dontrun{
rc <- sharing_ip_score()
}

}

