% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_scoring.R
\name{sharing_ips_with_employer_score}
\alias{sharing_ips_with_employer_score}
\title{Sharing IPs with employers}
\usage{
sharing_ips_with_employer_score(cycle_date = NULL, db_schema = "nrd",
  exclude_cellular = NULL)
}
\arguments{
\item{cycle_date}{The APM cycle date.}

\item{db_schema}{The database schema to write scores and related tables to.
Default: "nrd".}

\item{exclude_cellular}{Whether or not the module should exclude
records identified as originating from cellular IPs.}
}
\value{
An integer status code, like other models.
}
\description{
This module reads in the ip base table and determines which users have IPs
that have been shared with employers' IP addresses used when filing their
Quarterly Wage Reports.  It assigns a risk score based on the total number of
times that sharing has occurred and the percentage of time a claimant has
shared with an employer's IP. The final score is an equally weighted
combination of these two factors.  One thing to note is that a claimant can
legitimately file multiple claims in a single week in order to claim money
for previous weeks. If they use the same IP which matches with a single
employer, the claimant will be counted as sharing the ip 5x. This was left
as-is because claiming multiple weeks at once can be considered slightly
suspicious behavior, thus allowing the count to reflect that suspicion.
}
\details{
The following parameters are pulled from a parameter table to run this model:
- ${APM_CYCL_DT}: Kettle date insert (or passed as function arg)
}
\examples{
\dontrun{
rc <- sharing_ips_with_employer_score()
}

}

