% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{store_model_version}
\alias{store_model_version}
\title{Store the model version as a data object}
\usage{
store_model_version(model.version)
}
\arguments{
\item{model.version}{Character vector describing the models bundled
with the package.}
}
\description{
Store the model version as a data object
}
\author{
Tom Shafer
}

