% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{store_state_inputs}
\alias{store_state_inputs}
\title{Store state-specific model inputs and parameters}
\usage{
store_state_inputs(state_inputs)
}
\arguments{
\item{state_inputs}{A list of list objects, one for each model}
}
\value{
Saves a list object to APM/Data that will be used to train and score
models
}
\description{
Creates a list of lists. The outer list will be created for each state and 
will include a list for each model. The list for each model has two entries:
  1. Vector of input variables to use for that model
  2. List of model parameters
}
\author{
Daniel Brannock
}

