% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_vector_machine.R
\name{support_vector_machine_score}
\alias{support_vector_machine_score}
\title{Score Certifications with a DAPM Support Vector Machine Model}
\usage{
support_vector_machine_score(cycle_date = NULL, prevDays = 90,
  inputDF = NULL, scoreTableName = "nrd.enty_score")
}
\arguments{
\item{cycle_date}{The current cycle date for scoring.}

\item{prevDays}{Number of days to look backwards for a certification filed by
a claimant. Claimants who do not have a certification within this
range will not be scored, but claimants who do have a certification
within this range will have the most recent filing scored.}

\item{inputDF}{A data frame of certifications to score. If
\code{inputDF=NULL} (the default), this function scores certifications
from the database. This function  assumed the data was taken from an
SQL query like that in the source.}

\item{scoreTableName}{The fully-qualified table into which to place the
scores. If \code{scoreTableName=NULL}, return a data frame instead of
writing to the database.}
}
\value{
Returns 0 on success and an integer > 0 if an error occured. These
        integer codes should be unique.
}
\description{
This is a simpler, more opinionated scoring function than has been included
in DAPM previously, but this routine will extend more easily to multiple
states with different data.
}
\details{
Assumptions:

1. The input data (either from \code{aggr_claimt_benf} or \code{inputDF}) is
   raw (not transformed).
2. A Support Vector Machine model object already exists and can be lazy
   loaded. This object will contain:
    - A list of functions to transform the raw data
    - A list of values to be imputed in the case of missing data
    - A list of features to be included in the scoring

Given these assumptions, this function will:

1. Take in raw data 
2. Impute missing values
3. Subset down to ID vars and predictors
4. Transform the predictors
5. Score with the model
6. Return scores

The source of the raw data and the final destination of the scores are
controlled by input parameters.
}
\author{
Tom Shafer
}

