% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{svm.storage}
\alias{svm.storage}
\title{Support vector machines model object}
\format{See training/scoring functions.}
\usage{
svm.storage
}
\description{
\strong{This model object should never be put into version control.}
}
\keyword{datasets}

