% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_exists}
\alias{table_exists}
\title{Determine if a table exists in a schema}
\usage{
table_exists(conn, scoreTableName)
}
\arguments{
\item{conn}{the \code{conn} part of the \code{db_conn} value returned from \code{initialize_db}.}

\item{scoreTableName}{fully-qualified table to determine if it exists.}
}
\description{
\code{table_exists} Determine if a table exists in a schema.
}
\author{
Wayne Folta
}

