% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_inputs_dt}
\alias{transform_inputs_dt}
\title{Transform columns by reference within a data table}
\usage{
transform_inputs_dt(dt, transform.list)
}
\arguments{
\item{dt}{A data table.}

\item{transform.list}{A list for which keys are column names in the data
table and values are functions to apply.}
}
\value{
Returns an object with "error" in \code{class(obj)} if
        \code{tryCatch} detects a problem, otherwise a list of messages.
}
\description{
Transform columns by reference within a data table
}
\author{
Tom Shafer
}

