% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_to_tbl_cols}
\alias{vec_to_tbl_cols}
\title{Merge a list to a comma-separated string, prepending a table name}
\usage{
vec_to_tbl_cols(vec, tbl)
}
\arguments{
\item{vec}{Character vector of column names.}

\item{tbl}{Database table name.}
}
\value{
String containing comma-joined column names.
}
\description{
E.g. c("x", "y") -> "TBL.x, TBL.y"
}
\author{
Tom Shafer
}

