% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_scores}
\alias{write_scores}
\title{Write Scores}
\usage{
write_scores(claimt_ids, scores, conn, score_cd, cycle_date, scoreTableName,
  timestamp)
}
\arguments{
\item{claimt_ids}{vector of character claimant ids.}

\item{scores}{vector of numerical scores.}

\item{conn}{the \code{conn} part of the \code{db_conn} value returned from \code{initialize_db}.}

\item{score_cd}{the the character string code of this score type.}

\item{cycle_date}{the cycle date (numeric YYYYQ) of the current cycle.}

\item{scoreTableName}{fully-qualified table to which to write scores. This table must already exist.}

\item{timestamp}{character string timestamp of when this ran (more granular than Cycle Date).}
}
\description{
This function writes function scores out to a score table. The
\code{scoreTableName} defines where, and this table must already exist.
}
\details{
\code{write_scores} writes out scores to a score table.
}
\author{
Wayne Folta
}

