% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{ZIP Codes database}
\format{A data.table with three variables:
   1. zip_code: 5-digit ZIP code
   2. Lat:      Latitude coordinate
   3. Long:     Longitude coordinate}
\usage{
zipcodes
}
\description{
ZIP Codes database
}
\keyword{datasets}

